[ANDROSDK-516] Rename DIModule -> PackageDIModule
