+++ core/src/main/java/org/hisp/dhis/android/core/D2DIComponent.java
@@ -1,8 +1,8 @@
 package org.hisp.dhis.android.core;
 
 import org.hisp.dhis.android.core.arch.api.retrofit.APIClientDIModule;
-import org.hisp.dhis.android.core.category.CategoryDIModule;
-import org.hisp.dhis.android.core.common.CommonDIModule;
+import org.hisp.dhis.android.core.category.CategoryPackageDIModule;
+import org.hisp.dhis.android.core.common.CommonPackageDIModule;
 import org.hisp.dhis.android.core.common.GenericCallData;
 import org.hisp.dhis.android.core.data.database.DatabaseDIModule;
 import org.hisp.dhis.android.core.dataelement.DataElementPackageDIModule;
@@ -10,14 +10,14 @@ import org.hisp.dhis.android.core.dataset.DataSetPackageDIModule;
 import org.hisp.dhis.android.core.datavalue.DataValuePackageDIModule;
 import org.hisp.dhis.android.core.domain.aggregated.AggregatedModule;
 import org.hisp.dhis.android.core.indicator.IndicatorPackageDIModule;
-import org.hisp.dhis.android.core.maintenance.MaintenanceDIModule;
+import org.hisp.dhis.android.core.maintenance.MaintenancePackageDIModule;
 import org.hisp.dhis.android.core.option.OptionPackageDIModule;
 import org.hisp.dhis.android.core.period.PeriodPackageDIModule;
-import org.hisp.dhis.android.core.relationship.RelationshipDIModule;
-import org.hisp.dhis.android.core.resource.ResourceDIModule;
+import org.hisp.dhis.android.core.relationship.RelationshipPackageDIModule;
+import org.hisp.dhis.android.core.resource.ResourcePackageDIModule;
 import org.hisp.dhis.android.core.resource.ResourceHandler;
-import org.hisp.dhis.android.core.settings.SystemSettingDIModule;
-import org.hisp.dhis.android.core.systeminfo.SystemInfoDIModule;
+import org.hisp.dhis.android.core.settings.SystemSettingPackageDIModule;
+import org.hisp.dhis.android.core.systeminfo.SystemInfoPackageDIModule;
 import org.hisp.dhis.android.core.user.UserPackageDIModule;
 import org.hisp.dhis.android.core.wipe.WipeDIModule;
 import org.hisp.dhis.android.core.wipe.WipeModule;
@@ -28,23 +28,24 @@ import dagger.Component;
 
 @Singleton
 @Component(modules = {
-        DatabaseDIModule.class,
         APIClientDIModule.class,
-        CommonDIModule.class,
-        ResourceDIModule.class,
-        SystemInfoDIModule.class,
-        SystemSettingDIModule.class,
-        RelationshipDIModule.class,
-        CategoryDIModule.class,
-        DataSetPackageDIModule.class,
+        DatabaseDIModule.class,
+        WipeDIModule.class,
+
+        CategoryPackageDIModule.class,
+        CommonPackageDIModule.class,
         DataElementPackageDIModule.class,
+        DataSetPackageDIModule.class,
         DataValuePackageDIModule.class,
         IndicatorPackageDIModule.class,
+        MaintenancePackageDIModule.class,
         OptionPackageDIModule.class,
         PeriodPackageDIModule.class,
-        UserPackageDIModule.class,
-        MaintenanceDIModule.class,
-        WipeDIModule.class}
+        RelationshipPackageDIModule.class,
+        ResourcePackageDIModule.class,
+        SystemInfoPackageDIModule.class,
+        SystemSettingPackageDIModule.class,
+        UserPackageDIModule.class}
 )
 public interface D2DIComponent {
 
@@ -56,23 +57,24 @@ public interface D2DIComponent {
 
     @Component.Builder
     interface Builder {
-        Builder databaseDIModule(DatabaseDIModule databaseDIModule);
         Builder apiClientDIModule(APIClientDIModule apiClientDIModule);
-        Builder commonDIModule(CommonDIModule commonDIModule);
-        Builder resourceDIModule(ResourceDIModule resourceDIModule);
-        Builder systemInfoDIModule(SystemInfoDIModule systemInfoDIModule);
-        Builder systemSettingDIModule(SystemSettingDIModule systemSettingDIModule);
-        Builder relationshipDIModule(RelationshipDIModule relationshipDIModule);
-        Builder categoryDIModule(CategoryDIModule categoryDIModule);
-        Builder dataSetPackageDIModule(DataSetPackageDIModule dataSetPackageDIModule);
+        Builder databaseDIModule(DatabaseDIModule databaseDIModule);
+        Builder wipeDIModule(WipeDIModule wipeDIModule);
+
+        Builder categoryPackageDIModule(CategoryPackageDIModule categoryPackageDIModule);
+        Builder commonPackageDIModule(CommonPackageDIModule commonPackageDIModule);
         Builder dataElementPackageDIModule(DataElementPackageDIModule dataElementPackageDIModule);
+        Builder dataSetPackageDIModule(DataSetPackageDIModule dataSetPackageDIModule);
         Builder dataValuePackageDIModule(DataValuePackageDIModule dataValuePackageDIModule);
         Builder indicatorPackageDIModule(IndicatorPackageDIModule indicatorPackageDIModule);
+        Builder maintenancePackageDIModule(MaintenancePackageDIModule maintenancePackageDIModule);
         Builder optionPackageDIModule(OptionPackageDIModule optionPackageDIModule);
         Builder periodPackageDIModule(PeriodPackageDIModule periodPackageDIModule);
+        Builder relationshipDIModule(RelationshipPackageDIModule relationshipPackageDIModule);
+        Builder resourcePackageDIModule(ResourcePackageDIModule resourcePackageDIModule);
+        Builder systemInfoPackageDIModule(SystemInfoPackageDIModule systemInfoPackageDIModule);
+        Builder systemSettingPackageDIModule(SystemSettingPackageDIModule systemSettingPackageDIModule);
         Builder userPackageDIModule(UserPackageDIModule userPackageDIModule);
-        Builder maintenanceDIModule(MaintenanceDIModule maintenanceDIModule);
-        Builder wipeDIModule(WipeDIModule wipeDIModule);
         D2DIComponent build();
     }
 }
